Bug fix: crashes when reading damaged images, caused by buffer overruns, or by deleting uninitialized pointers (found by Apple).

+++ OpenEXR/IlmImf/ImfB44Compressor.h
@@ -54,8 +54,8 @@ class B44Compressor: public Compressor
   public:
 
     B44Compressor (const Header &hdr,
-                   int maxScanLineSize,
-		   int numScanLines,
+                   size_t maxScanLineSize,
+		   size_t numScanLines,
 		   bool optFlatFields);
 
     virtual ~B44Compressor ();
